Cleaned up error handeling
Error handeling can now be down with fail() which will log the failure,
and update the status of instance with FAILED.  It it will ignore update
failures because of a more important error.

The StackCreateFailed status is treated just like other Failed status.
This status will also leave a stack that has to be deleted during
cleanup.  It should not have the finalizers removed.