EKS will now delete components with label "eks.needsdeleting":"true", (#45)
* EKS will now delete components with label "eks.needsdeleting":"true",

Components defined in the pkg/controller/eks/components.go will now be
deleted if they have the correct label.  This will happen before
removing the nodegroups or cluster to allow for any remote
operators to clean up any existing resources.

Components were made to stick around until the remote
object is deleted.

* Added test to confirm components get deleted.

* Added components finalizer to the default eks instances.

This is done to force a check of components before deleting the cluster.